Remove blockchain porter tests and flatten test structure. Update dependencies to use `pytest-nucypher` plugin.
