+++ dev-requirements.txt
@@ -1,7 +1,7 @@
 -i https://pypi.python.org/simple
 attrs==22.1.0; python_version >= '3.5'
 cfgv==3.3.1; python_full_version >= '3.6.1'
-coverage==6.5.0
+coverage[toml]==6.5.0
 distlib==0.3.6
 exceptiongroup==1.0.0; python_version < '3.11'
 filelock==3.8.0; python_version >= '3.7'
@@ -20,8 +20,7 @@ pytest-cov==4.0.0
 pytest-mock==3.10.0
 pytest==7.2.0
 pyyaml==6.0; python_version >= '3.6'
-setuptools==65.4.1; python_version >= '3.7'
 toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
-tomli==2.0.1; python_version >= '3.7'
+tomli==2.0.1; python_version >= '3.7' and python_version < '3.11'
 typing-extensions==4.3.0; python_version >= '3.7'
 virtualenv==20.16.6; python_version >= '3.6'
