+++ requirements.txt
@@ -1,5 +1,4 @@
 -i https://pypi.python.org/simple
-git+https://github.com/nucypher/nucypher.git@919d6b45c28523a2ff08059e503a7a2bca04de81#egg=nucypher
 aiohttp==3.8.3; python_version >= '3.6'
 aiosignal==1.2.0; python_version >= '3.6'
 appdirs==1.4.4
@@ -11,10 +10,8 @@ base58==2.1.1; python_version >= '3.5'
 bitarray==2.6.0
 bytestring-splitter==2.4.1
 cached-property==1.5.2
-cachetools==5.2.0; python_version ~= '3.7'
 certifi==2022.9.24; python_version >= '3.6'
 cffi==1.15.1
-chardet==5.0.0; python_version >= '3.6'
 charset-normalizer==2.1.1; python_version >= '3.6'
 click==8.1.3; python_version >= '3.7'
 colorama==0.4.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
@@ -22,7 +19,7 @@ constant-sorrow==0.1.0a9; python_version >= '3'
 constantly==15.1.0
 construct==2.10.68; python_version >= '3.6'
 cryptography==38.0.1; python_version >= '3.6'
-cytoolz==0.12.0; python_version >= '3.5'
+cytoolz==0.12.0; implementation_name == 'cpython'
 dateparser==1.1.1; python_version >= '3.5'
 distlib==0.3.6
 ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
@@ -30,7 +27,7 @@ eip712-structs==1.1.0
 eth-abi==3.0.1; python_version >= '3.7' and python_version < '4'
 eth-account==0.7.0; python_version >= '3.6' and python_version < '4'
 eth-bloom==1.0.4; python_version >= '3.6' and python_version < '4'
-eth-hash==0.3.3; python_version >= '3.5' and python_version < '4'
+eth-hash[pycryptodome]==0.3.3; python_version >= '3.5' and python_version < '4'
 eth-keyfile==0.6.0
 eth-keys==0.4.0
 eth-rlp==0.3.0; python_version >= '3.7' and python_version < '4'
@@ -42,11 +39,13 @@ flask-cors==3.0.10
 flask-htpasswd==0.5.0
 flask==2.2.2; python_version >= '3.7'
 frozenlist==1.3.1; python_version >= '3.7'
+git+https://github.com/nucypher/nucypher.git@2403f9f765654ff74327127e7219ecd909a8eeb2#egg=nucypher
 hendrix==3.4.0
 hexbytes==0.3.0; python_version >= '3.7' and python_version < '4'
 humanize==4.4.0; python_version >= '3.7'
 hyperlink==21.0.0
 idna==3.4; python_version >= '3.5'
+importlib-metadata==5.0.0; python_version < '3.10'
 incremental==21.3.0
 ipfshttpclient==0.8.0a2; python_full_version >= '3.6.2' and python_full_version not in '3.7.0, 3.7.1'
 itsdangerous==2.0.1; python_version >= '3.6'
@@ -77,17 +76,17 @@ pluggy==1.0.0; python_version >= '3.6'
 protobuf==3.20.1; python_version >= '3.7'
 py-ecc==6.0.0; python_version >= '3.6' and python_version < '4'
 py-evm==0.6.0a1
+py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
 pyasn1-modules==0.3.0rc1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
 pyasn1==0.5.0rc2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
 pychalk==2.0.1
 pycparser==2.21
-pycryptodome==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+pycryptodome==3.15.0
 pyethash==0.1.27
 pyjwt==2.6.0; python_version >= '3.7'
 pynacl==1.5.0; python_version >= '3.6'
 pyopenssl==22.1.0; python_version >= '3.6'
 pyparsing==3.0.9; python_full_version >= '3.6.8'
-pyproject-api==1.1.1; python_version >= '3.7'
 pyrsistent==0.18.1; python_version >= '3.7'
 pysha3==1.0.2
 python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
@@ -99,15 +98,14 @@ requests==2.28.1; python_version >= '3.7' and python_version < '4'
 rlp==3.0.0
 semantic-version==2.10.0; python_version >= '2.7'
 service-identity==21.1.0
-setuptools==65.4.1; python_version >= '3.7'
-simple-rlp==0.1.3; python_version >= '3.6'
+simple-rlp==0.1.3; python_version >= '3.7' and python_version >= '3.7'
 six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
 snaptime==0.2.4
 sortedcontainers==2.4.0
 tabulate==0.8.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-tomli==2.0.1; python_version >= '3.7'
+tomli==2.0.1; python_version >= '3.7' and python_version < '3.11'
 toolz==0.12.0; python_version >= '3.5'
-tox==4.0.0b2; python_version >= '3.7'
+tox==3.27.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
 trezor==0.13.3; python_version >= '3.6'
 trie==2.0.2; python_version >= '3.6' and python_version < '4'
 twisted==22.8.0; python_full_version >= '3.7.1'
@@ -118,8 +116,9 @@ urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3
 varint==1.0.2
 virtualenv==20.16.6; python_version >= '3.6'
 watchdog==2.1.9; python_version >= '3.6'
-web3==6.0.0b6
+web3==6.0.0b6; python_full_version >= '3.7.2'
 websockets==10.3; python_version >= '3.7'
 werkzeug==2.2.2; python_version >= '3.7'
 yarl==1.8.1; python_version >= '3.7'
+zipp==3.10.0; python_version >= '3.7'
 zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
